+++ R/file_client_funcs.R
@@ -208,7 +208,7 @@ delete_file_share.file_endpoint <- function(endpoint, name, confirm=TRUE, ...)
 #' @param info Whether to return names only, or all information in a directory listing.
 #' @param src,dest The source and destination files for uploading and downloading. See 'Details' below.
 #' @param confirm Whether to ask for confirmation on deleting a file or directory.
-#' @param recursive For the multiupload/download functions, whether to recursively transfer files in subdirectories. For `list_azure_dir`, whether to include the contents of any subdirectories in the listing. For `create_azure_dir`, whether to recursively create each component of a nested directory path. For `delete_azure_dir`, whether to delete a subdirectory's contents first (not yet supported). Note that in all cases this can be slow, so try to use a non-recursive solution if possible.
+#' @param recursive For the multiupload/download functions, whether to recursively transfer files in subdirectories. For `list_azure_dir`, whether to include the contents of any subdirectories in the listing. For `create_azure_dir`, whether to recursively create each component of a nested directory path. For `delete_azure_dir`, whether to delete a subdirectory's contents first. Note that in all cases this can be slow, so try to use a non-recursive solution if possible.
 #' @param create_dir For the uploading functions, whether to create the destination directory if it doesn't exist. Again for the file storage API this can be slow, hence is optional.
 #' @param blocksize The number of bytes to upload/download per HTTP(S) request.
 #' @param overwrite When downloading, whether to overwrite an existing destination file.
@@ -308,7 +308,7 @@ list_azure_files <- function(share, dir="/", info=c("all", "name"),
     }
 
     name <- vapply(out, function(ent) ent$Name[[1]], FUN.VALUE=character(1))
-    isdir <- if(is_empty(name)) character(0) else names(name) == "Directory"
+    isdir <- if(is_empty(name)) logical(0) else names(name) == "Directory"
     size <- vapply(out,
                    function(ent) if(is_empty(ent$Properties)) NA_character_
                                  else ent$Properties$`Content-Length`[[1]],
@@ -322,7 +322,7 @@ list_azure_files <- function(share, dir="/", info=c("all", "name"),
         dirs <- df$name[df$isdir]
 
         nextlevel <- lapply(dirs, function(d) list_azure_files(share, d, info="all", prefix=prefix, recursive=TRUE))
-        df <- do.call(rbind, c(list(df), nextlevel))
+        df <- do.call(vctrs::vec_rbind, c(list(df), nextlevel))
     }
 
     if(info == "name")
@@ -405,15 +405,27 @@ create_azure_dir <- function(share, dir, recursive=FALSE)
 #' @export
 delete_azure_dir <- function(share, dir, recursive=FALSE, confirm=TRUE)
 {
-    if(dir %in% c("/", "."))
+    if(dir %in% c("/", ".") && !recursive)
         return(invisible(NULL))
 
     if(!delete_confirmed(confirm, paste0(share$endpoint$url, share$name, "/", dir), "directory"))
         return(invisible(NULL))
 
     if(recursive)
-        stop("Recursive deleting of subdirectory contents not yet supported", call.=FALSE)
+    {
+        conts <- list_azure_files(share, dir, recursive=TRUE)
+        for(i in rev(seq_len(nrow(conts))))
+        {
+            # delete all files and dirs
+            # assumption is that files will be listed after their parent dir
+            if(conts$isdir[i])
+                delete_azure_dir(share, conts$name[i], recursive=FALSE, confirm=FALSE)
+            else delete_azure_file(share, conts$name[i], confirm=FALSE)
+        }
+    }
 
+    if(dir == "/")
+        return(invisible(NULL))
     invisible(do_container_op(share, dir, options=list(restype="directory"), http_verb="DELETE"))
 }