+++ NEWS.md
@@ -1,3 +1,10 @@
+# AzureStor 3.5.2
+
+- Use the vctrs package to parse paged output from the REST API. This should resolve many issues with `list_blobs` in particular.
+- Add support for recursive deleting of non-empty subdirectories in blob and file storage.
+- Fix a bug when listing files in an empty ADLS filesystem (#100).
+- Fix passing a timeout parameter to the REST API (#105).
+
 # AzureStor 3.5.1
 
 - Fix for changed behaviour in readr 2.0.0, where `readr::read_delim` now automatically closes an input connection (#95).